small-refactoring. move internal independent classes to separate class file.
