+++ soapui/src/main/java/com/eviware/soapui/model/tree/nodes/PropertiesTreeNode.java
@@ -1,17 +1,17 @@
 /*
  * SoapUI, Copyright (C) 2004-2019 SmartBear Software
  *
- * Licensed under the EUPL, Version 1.1 or - as soon as they will be approved by the European Commission - subsequent 
- * versions of the EUPL (the "Licence"); 
- * You may not use this work except in compliance with the Licence. 
- * You may obtain a copy of the Licence at: 
- * 
- * http://ec.europa.eu/idabc/eupl 
- * 
- * Unless required by applicable law or agreed to in writing, software distributed under the Licence is 
- * distributed on an "AS IS" basis, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either 
- * express or implied. See the Licence for the specific language governing permissions and limitations 
- * under the Licence. 
+ * Licensed under the EUPL, Version 1.1 or - as soon as they will be approved by the European Commission - subsequent
+ * versions of the EUPL (the "Licence");
+ * You may not use this work except in compliance with the Licence.
+ * You may obtain a copy of the Licence at:
+ *
+ * http://ec.europa.eu/idabc/eupl
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the Licence is
+ * distributed on an "AS IS" basis, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the Licence for the specific language governing permissions and limitations
+ * under the Licence.
  */
 
 package com.eviware.soapui.model.tree.nodes;
@@ -24,7 +24,6 @@ import com.eviware.soapui.model.testsuite.TestPropertyListener;
 import com.eviware.soapui.model.tree.AbstractModelItemTreeNode;
 import com.eviware.soapui.model.tree.SoapUITreeModel;
 import com.eviware.soapui.model.tree.SoapUITreeNode;
-import com.eviware.soapui.model.tree.nodes.support.EmptyModelItem;
 import com.eviware.soapui.support.StringUtils;
 import com.eviware.soapui.support.UISupport;
 import com.eviware.soapui.support.action.swing.ActionList;
@@ -61,8 +60,8 @@ public class PropertiesTreeNode<T extends ModelItem> extends AbstractModelItemTr
     }
 
     public static PropertiesTreeNode<?> createDefaultPropertiesNode(TestModelItem modelItem, SoapUITreeModel treeModel) {
-        return new PropertiesTreeNode<PropertiesTreeNode.PropertiesModelItem>(
-                new PropertiesTreeNode.PropertiesModelItem(modelItem), modelItem, modelItem, treeModel);
+        return new PropertiesTreeNode<PropertiesModelItem>(
+                new PropertiesModelItem(modelItem), modelItem, modelItem, treeModel);
     }
 
     public int getChildCount() {
@@ -143,20 +142,6 @@ public class PropertiesTreeNode<T extends ModelItem> extends AbstractModelItemTr
             }
 
             getTreeModel().notifyNodeInserted(node);
-        }
-    }
-
-    public static class PropertiesModelItem extends EmptyModelItem {
-        private final TestPropertyHolder holder;
-
-        public PropertiesModelItem(TestPropertyHolder holder) {
-            super("Properties (" + holder.getPropertyNames().length + ")", UISupport
-                    .createImageIcon("/properties_step.png"));
-            this.holder = holder;
-        }
-
-        public void updateName() {
-            setName("Properties (" + holder.getPropertyNames().length + ")");
         }
     }
