+++ soapui/src/main/java/com/eviware/soapui/support/propertyexpansion/PropertyExpansionDropTarget.java
@@ -19,7 +19,7 @@ package com.eviware.soapui.support.propertyexpansion;
 import com.eviware.soapui.SoapUI;
 import com.eviware.soapui.model.propertyexpansion.PropertyExpansionImpl;
 import com.eviware.soapui.model.propertyexpansion.PropertyExpansionUtils;
-import com.eviware.soapui.model.tree.nodes.PropertyTreeNode.PropertyModelItem;
+import com.eviware.soapui.model.tree.nodes.PropertyModelItem;
 import com.eviware.soapui.support.UISupport;
 import com.eviware.soapui.support.xml.XmlUtils;
